Added live color change for whole strip
In qstrip we now have new button called "UniColor".

This button opens a QColorDialog, which lets you pick a color. This color is directly applied to the whole strip (i.e. the currently set number of LEDs).

If you cancel the dialog the previous state (at the moment of opening the dialog) is restored.